% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{coverage_prob}
\alias{coverage_prob}
\title{Calculate the empirical coverage probability of a matrix of confidence
intervals.}
\usage{
coverage_prob(conf_intervals, truth)
}
\arguments{
\item{conf_intervals}{Matrix of confidence intervals in which each row is a
CI, the first column is the lower confidence limit, and the second column is
the upper confidence limit.}

\item{truth}{True value of the parameter for which the CIs were constructed.}
}
\value{
Empirical coverage probability.
}
\description{
Calculate the empirical coverage probability of a matrix of confidence
intervals.
}
\examples{
xbar <- replicate(1000, mean(rnorm(25)))
ME <- qnorm(0.975) / 5
CIs <- cbind(xbar - ME, xbar + ME)
coverage_prob(CIs, 0)
}

