% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAD}
\alias{MAD}
\title{Calculate median absolute deviation}
\usage{
MAD(x, truth)
}
\arguments{
\item{x}{Vector of estimates.}

\item{truth}{True value of the parameter.}
}
\value{
Median absolute deviation.
}
\description{
Calculate median absolute deviation
}
\examples{
x <- rnorm(1000) + 0.5
MAD(x, 0)
}

