% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{median_width}
\alias{median_width}
\title{Calculate empirical median width of a matrix of confidence intervals.}
\usage{
median_width(conf_intervals)
}
\arguments{
\item{conf_intervals}{Matrix of confidence intervals in which each row is a
CI, the first column is the lower confidence limit, and the second column is
the upper confidence limit.}
}
\value{
Empirical median width of the confidence intervals.
}
\description{
Calculate empirical median width of a matrix of confidence intervals.
}
\examples{
xbar <- replicate(1000, mean(rnorm(25)))
ME <- qnorm(0.975) / 5
CIs <- cbind(xbar - ME, xbar + ME)
median_width(CIs)
}

