% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Generate draws from a multivariate normal distribution}
\usage{
mvrnorm(n, mu, Sigma)
}
\arguments{
\item{n}{Number of samples.}

\item{mu}{Mean vector.}

\item{Sigma}{Covariance matrix.}
}
\value{
Draws from the normal distribution.
}
\description{
Generate draws from a multivariate normal distribution
}
\details{
This is essentially a stripped-down version of the mvrnorm function
from the MASS library in R. Through the magic of Rcpp we're transforming the
same standard normal draws as the R version. However, since Armadillo
follows a different convention from R in its definition of the
eign-decomposition, the output of this function will *not* be the same as
that of its R counterpart. Since we access R's function for generating
normal draws, we can set the seed from R.
}
\examples{
mvrnorm(10, c(0,0), diag(1, 2, 2))
}

