% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLSvsIV_limit_sim}
\alias{OLSvsIV_limit_sim}
\title{Testing interface to limit_sim_OLS_IV class.}
\usage{
OLSvsIV_limit_sim(tau, pi_sq, n_sim = 10000L)
}
\arguments{
\item{tau}{Controls degree of endogeneity of OLS.}

\item{pi_sq}{First-stage R-squared (strengh of instruments).}

\item{n_sim}{Number of simulation draws.}
}
\value{
Dataframe of draws from the limit experiment for OLS, TSLS and the
sample estimator of tau.
}
\description{
Testing interface to limit_sim_OLS_IV class.
}
\details{
This is a testing interface to the C++ class that is used to draw
from the limit experiment in the OLS versus TSLS simulation experiment from
section 5.1 of the paper.
}
\examples{
foo <- OLSvsIV_limit_sim(5, 0.1)
cov(data.frame(ols = foo$ols, tsls = foo$tsls, tauhat = foo$tauhat))
}

