% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLSvsIV_nonsimCI}
\alias{OLSvsIV_nonsimCI}
\title{Coverage and Width of OLS, TSLS and Naive CIs}
\usage{
OLSvsIV_nonsimCI(tau, pi_sq, size = 0.05, n_sim = 10000L)
}
\arguments{
\item{tau}{Controls degree of endogeneity of OLS.}

\item{pi_sq}{First-stage R-squared (strengh of instruments).}

\item{size}{One minus the norminal coverage probability of the intervals.}

\item{n_sim}{Number of simulation draws.}
}
\value{
List containing empirical coverage probabilities and median width
of confidence intervals for the OLS and TSLS estimators and the same for a
"naive" confidence interval for the FMSC-selected estimator.
}
\description{
Coverage and Width of OLS, TSLS and Naive CIs
}
\details{
This function gives results based on simulations from the limit
experiment for the OLS versus TSLS example in Section 5.1 of the paper.
The confidence intervals computed here are for the non-simulation based
procedures: the OLS estimator, the TSLS estimator and a naive interval for
the post-FMSC estimator. This naive interval is constructed from the
textbook interval for whichever estimator the FMSC selects: if OLS is
selected it uses the standard OLS interval, and if TSLS is selected it uses
the TSLS interval. This procedure can perform very badly depending on
parameter values. Note that the median widths in this example are not
particularly interesting: the width of each OLS and TSLS interval is fixed
across all simulations and the median width of the naive interval equals
that of OLS when OLS is chosen more than 50 percent of the time and equals
that of TSLS otherwise. The median widths are provided merely for
consistency with other functions for which this quantity is more
interesting, namely the simulation-based intervals that try to correct some
of the deficiencies of the naive interval.
}
\examples{
foo <- OLSvsIV_nonsimCI(tau = 3, pi_sq = 0.1)
as.data.fram(foo)
}

