% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{OLSvsIV_second_step}
\alias{OLSvsIV_second_step}
\title{Testing interface to second_step_OLS_IV class.}
\usage{
OLSvsIV_second_step(tau, pi_sq, size = 0.05, inc = 0.001, n_sim = 1000L)
}
\arguments{
\item{tau}{Controls degree of endogeneity of OLS.}

\item{pi_sq}{First-stage R-squared (strengh of instruments).}

\item{size}{One minus the norminal coverage probability of the intervals.}

\item{inc}{Step size for grid over which width is minimized.}

\item{n_sim}{Number of simulation draws.}
}
\value{
List with two elements: simulation-based two-sided confidence
interval for the post-FMSC estimator evaluated at tau and the same for the
moment average estimator.
}
\description{
Testing interface to second_step_OLS_IV class.
}
\details{
This is a testing interface to the C++ class that is use, among
other things, as the second-step in the two-step simulation based confidence
interval construction from Section 4.3 of the paper, based on drawing from
the limit experiment. In this implementation it treats all population
parameters that enter the limit as know with the exception of tau. (That is,
it abstracts from sampling uncertainty.)
}
\examples{
as.data.frame(OLSvsIV_second_step(tau = 3, pi_sq = 0.1))
}

