% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_quantile}
\alias{sample_quantile}
\title{Calculate a sample quantile}
\usage{
sample_quantile(x, p)
}
\arguments{
\item{x}{Vector of data.}

\item{p}{Probability for desired quantile (e.g. 0.5 for median)}
}
\value{
Sample quantile
}
\description{
Calculate a sample quantile
}
\details{
There are many competing definitions of sample quantiles
(see Hyndman & Fan, 1996). Here we simply use the R default definition,
which corresponds to Definition 7 in Hyndman & Fan. See ?quantile in R for
more details.
}
\examples{
foo <- rnorm(1000)
sample_quantile(foo, 0.16)
}

