% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{shortest_CI}
\alias{shortest_CI}
\title{Calculate shortest two-sided confidence interval}
\usage{
shortest_CI(x, size = 0.05, inc = 0.001)
}
\arguments{
\item{x}{Vector of simulations from sampling distribution of an estimator.}

\item{size}{One minus the desired coverage probability.}

\item{inc}{Step size for grid over which width is minimized.}
}
\value{
Shortest (1 - size) * 100 percent interval based on the data provided.
}
\description{
Calculate shortest two-sided confidence interval
}
\examples{
x <- rnorm(1000)
shortest_CI(x)
}

